Codes for weather types

This data frame consists of the weather description codes used in the ISD. It is not of general use to most users.

Usage

weatherCodes

Format

An object of class data.frame with 100 rows and 2 columns.

Details

pwc

Weather code, which can be merged with the pwc column in importNOAA() datasets.

description

Description associated with the weather codes.

Examples

weatherCodes
#>     pwc
#> 1    00
#> 2    01
#> 3    02
#> 4    03
#> 5    04
#> 6    05
#> 7    06
#> 8    07
#> 9    08
#> 10   09
#> 11   10
#> 12   11
#> 13   12
#> 14   13
#> 15   14
#> 16   15
#> 17   16
#> 18   17
#> 19   18
#> 20   19
#> 21   20
#> 22   21
#> 23   22
#> 24   23
#> 25   24
#> 26   25
#> 27   26
#> 28   27
#> 29   28
#> 30   29
#> 31   30
#> 32   31
#> 33   32
#> 34   33
#> 35   34
#> 36   35
#> 37   36
#> 38   37
#> 39   38
#> 40   39
#> 41   40
#> 42   41
#> 43   42
#> 44   43
#> 45   44
#> 46   45
#> 47   46
#> 48   47
#> 49   48
#> 50   49
#> 51   50
#> 52   51
#> 53   52
#> 54   53
#> 55   54
#> 56   55
#> 57   56
#> 58   57
#> 59   58
#> 60   59
#> 61   60
#> 62   61
#> 63   62
#> 64   63
#> 65   64
#> 66   65
#> 67   66
#> 68   67
#> 69   68
#> 70   69
#> 71   70
#> 72   71
#> 73   72
#> 74   73
#> 75   74
#> 76   75
#> 77   76
#> 78   77
#> 79   78
#> 80   79
#> 81   80
#> 82   81
#> 83   82
#> 84   83
#> 85   84
#> 86   85
#> 87   86
#> 88   87
#> 89   88
#> 90   89
#> 91   90
#> 92   91
#> 93   92
#> 94   93
#> 95   94
#> 96   95
#> 97   96
#> 98   97
#> 99   98
#> 100  99
#>                                                                    description
#> 1                                                                  clear skies
#> 2                                                            clouds dissolving
#> 3                                                       state of sky unchanged
#> 4                                                            clouds developing
#> 5                                                  visibility reduced by smoke
#> 6                                                                        smoke
#> 7                             widespread dust in suspension not raised by wind
#> 8                                                  dust or sand raised by wind
#> 9                                           well developed dust or sand whirls
#> 10                          dust or sand storm within sight but not at station
#> 11                                                                        mist
#> 12                                                                diamond dust
#> 13                                                           distant lightning
#> 14                                         lightning visible, no thunder heard
#> 15                           precipitation within sight but not hitting ground
#> 16                            distant precipitation but not falling at station
#> 17                             nearby precipitation but not falling at station
#> 18                        thunderstorm but no precipitation falling at station
#> 19                squalls within sight but no precipitation falling at station
#> 20                                                  funnel clouds within sight
#> 21                                                                         fog
#> 22                                                                        rain
#> 23                                       drizzle (not freezing) or snow grains
#> 24                                                                        rain
#> 25                                                                        snow
#> 26                                           freezing drizzle or freezing rain
#> 27                               thunderstorms (with or without precipitation)
#> 28                                            blowing or drifting snow or sand
#> 29  blowing or drifting snow or sand, visibility equal to or greater than 1 km
#> 30                 blowing or drifting snow or sand, visibility less than 1 km
#> 31                                                                         fog
#> 32                                                      fog or ice fog patches
#> 33                     fog or ice fog, has become thinner during the past hour
#> 34                  fog or ice fog, no appreciable change during the past hour
#> 35                     fog or ice fog, has become thicker during the past hour
#> 36                                                        fog, depositing rime
#> 37                                                                        <NA>
#> 38                                                                        <NA>
#> 39                                                                        <NA>
#> 40                                                                        <NA>
#> 41                                                               precipitation
#> 42                                           precipitation, slight or moderate
#> 43                                                        precipitation, heavy
#> 44                                    liquid precipitation, slight or moderate
#> 45                                                 liquid precipitation, heavy
#> 46                                     solid precipitation, slight or moderate
#> 47                                                  solid precipitation, heavy
#> 48                                  freezing precipitation, slight or moderate
#> 49                                               freezing precipitation, heavy
#> 50                                                                        <NA>
#> 51                                                                     drizzle
#> 52                                               drizzle, not freezing, slight
#> 53                                             drizzle, not freezing, moderate
#> 54                                                drizzle, not freezing, heavy
#> 55                                                   drizzle, freezing, slight
#> 56                                                 drizzle, freezing, moderate
#> 57                                                    drizzle, freezing, heavy
#> 58                                                    drizzle and rain, slight
#> 59                                         drizzle and rain, moderate or heavy
#> 60                                                                        <NA>
#> 61                                                                        rain
#> 62                                                  rain, not freezing, slight
#> 63                                                rain, not freezing, moderate
#> 64                                                   rain, not freezing, heavy
#> 65                                                      rain, freezing, slight
#> 66                                                    rain, freezing, moderate
#> 67                                                       rain, freezing, heavy
#> 68                                            rain or drizzle and snow, slight
#> 69                                 rain or drizzle and snow, moderate or heavy
#> 70                                                                        <NA>
#> 71                                                                        snow
#> 72                                                                snow, slight
#> 73                                                              snow, moderate
#> 74                                                                 snow, heavy
#> 75                                                         ice pellets, slight
#> 76                                                       ice pellets, moderate
#> 77                                                          ice pellets, heavy
#> 78                                                                 snow grains
#> 79                                                                ice crystals
#> 80                                                                        <NA>
#> 81                                       showers or intermittent precipitation
#> 82                                   rain showers or intermittent rain, slight
#> 83                                 rain showers or intermittent rain, moderate
#> 84                                    rain showers or intermittent rain, heavy
#> 85                                  rain showers or intermittent rain, violent
#> 86                                   snow showers or intermittent snow, slight
#> 87                                 snow showers or intermittent snow, moderate
#> 88                                    snow showers or intermittent snow, heavy
#> 89                                                                        <NA>
#> 90                                                                        hail
#> 91                                                                thunderstorm
#> 92                     thunderstorm, slight or moderate, with no precipitation
#> 93     thunderstorm, slight or moderate, with rain showers and/or snow showers
#> 94                                 thunderstorm, slight or moderate, with hail
#> 95                                   thunderstorm, heavy with no precipitation
#> 96                           thunderstorm, heavy with rain showers and/or snow
#> 97                                               thunderstorm, heavy with hail
#> 98                                                                        <NA>
#> 99                                                                        <NA>
#> 100                                                                    tornado